requirements.txt.txt
django==4
python-docx
bs4
scikit_learn==0.24.2
matplotlib
pandas
nltk